# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Orthoplex API",
        "description": "A Laravel 12 multi-tenant SaaS API with comprehensive authentication, RBAC, and user management.",
        "contact": {
            "email": "support@orthoplex.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://orthoplex.test",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/analytics/overview": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get login analytics overview",
                "description": "Get comprehensive login analytics including totals, trends, and distributions",
                "operationId": "af8e066b45128ed3108442db0aef22f2",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Analytics period",
                        "schema": {
                            "type": "string",
                            "default": "30d",
                            "enum": [
                                "7d",
                                "30d",
                                "90d",
                                "1y"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analytics data retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_logins": {
                                            "type": "integer",
                                            "example": 1250
                                        },
                                        "unique_users": {
                                            "type": "integer",
                                            "example": 89
                                        },
                                        "successful_logins": {
                                            "type": "integer",
                                            "example": 1200
                                        },
                                        "failed_logins": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "success_rate": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 96
                                        },
                                        "daily_trends": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "date": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "2024-01-15"
                                                    },
                                                    "logins": {
                                                        "type": "integer",
                                                        "example": 45
                                                    },
                                                    "unique_users": {
                                                        "type": "integer",
                                                        "example": 12
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "method_distribution": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "method": {
                                                        "type": "string",
                                                        "example": "password"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 800
                                                    },
                                                    "percentage": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 64
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "geographic_distribution": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "country": {
                                                        "type": "string",
                                                        "example": "United States"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 450
                                                    },
                                                    "percentage": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 36
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/analytics/security": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get security analytics",
                "description": "Get security-focused analytics including failed login attempts, suspicious activity",
                "operationId": "c403c87ec55aa3180667ab3051061d95",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Analytics period",
                        "schema": {
                            "type": "string",
                            "default": "30d",
                            "enum": [
                                "7d",
                                "30d",
                                "90d"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Security analytics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "total_failed_attempts": {
                                            "type": "integer",
                                            "example": 125
                                        },
                                        "rate_limited_attempts": {
                                            "type": "integer",
                                            "example": 45
                                        },
                                        "suspicious_ips": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "ip_address": {
                                                        "type": "string",
                                                        "example": "192.168.1.100"
                                                    },
                                                    "failed_attempts": {
                                                        "type": "integer",
                                                        "example": 15
                                                    },
                                                    "last_attempt": {
                                                        "type": "string",
                                                        "format": "datetime"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "failed_by_reason": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "reason": {
                                                        "type": "string",
                                                        "example": "invalid_credentials"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 80
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/analytics/users": {
            "get": {
                "tags": [
                    "Analytics"
                ],
                "summary": "Get user behavior analytics",
                "description": "Get user-focused analytics including activity patterns, device usage",
                "operationId": "ac37646dec0ce2cfa52bef070ec6fa40",
                "parameters": [
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Analytics period",
                        "schema": {
                            "type": "string",
                            "default": "30d",
                            "enum": [
                                "7d",
                                "30d",
                                "90d"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User analytics retrieved successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}